ISPN-9922 put search semantics in KeyValueStore API
