--- api/src/main/java/org/infinispan/api/search/reactive/SearchableStore.java
@@ -1,7 +1,0 @@
-package org.infinispan.api.search.reactive;
-
-public interface SearchableStore<V> {
-   ReactiveQuery find(String ickleQuery);
-
-   ReactiveContinuousQuery findContinuously(String ickleQuery);
-}
